+++ sandbox/auth-ms-basic-example/src/enums/permission.enum.ts
@@ -1,0 +1,5 @@
+export declare const enum PermissionKey {
+  CreateTodo = 'CreateTodo',
+  UpdateTodo = 'UpdateTodo',
+  DeleteTodo = 'DeleteTodo',
+}