+++ sandbox/auth-ms-basic-example/src/index.ts
@@ -1,0 +1,51 @@
+import dotenv from 'dotenv';
+import dotenvExt from 'dotenv-extended';
+import {
+  ApplicationConfig,
+  AuthMultitenantExampleApplication,
+} from './application';
+export * from './application';
+
+dotenv.config();
+dotenvExt.load({
+  schema: '.env.example',
+  errorOnMissing: true,
+  includeProcessEnv: true,
+});
+const PORT = 3000;
+
+export async function main(options: ApplicationConfig = {}) {
+  const app = new AuthMultitenantExampleApplication(options);
+  await app.boot();
+  await app.start();
+
+  const url = app.restServer.url;
+  console.log(`Server is running at ${url}`);
+  console.log(`Try ${url}/ping`);
+
+  return app;
+}
+
+if (require.main === module) {
+  // Run the application
+  const config = {
+    rest: {
+      port: +(process.env.PORT ?? PORT),
+      host: process.env.HOST,
+      // The `gracePeriodForClose` provides a graceful close for http/https
+      // servers with keep-alive clients. The default value is `Infinity`
+      // (don't force-close). If you want to immediately destroy all sockets
+      // upon stop, set its value to `0`.
+      // See https://www.npmjs.com/package/stoppable
+      gracePeriodForClose: 5000, // 5 seconds
+      openApiSpec: {
+        // useful when used with OpenAPI-to-GraphQL to locate your application
+        setServersFromRequest: true,
+      },
+    },
+  };
+  main(config).catch(err => {
+    console.error('Cannot start the application.', err);
+    process.exit(1);
+  });
+}