+++ sandbox/auth-ms-basic-example/src/models/index.ts
@@ -1,0 +1,2 @@
+export * from './to-do.model';
+export * from './user.dto';